Update species_model_v3.R
Kiona made lots of changes to code, mostly to fix and expand upon the parts associated with the hierarchical importance weights, and to compute various quantities at the population (species) level. Also note that I expanded upon the "model" for the AR1 data, which introduced some new parameters, and initials will need to be provided for:
mu.mu.AR1
sig.AR1
sig.mu.AR1
The data for Status and SiteID and will need to be updated so that they are vectors of length "Ncores" so that they give the Status and SiteID for a given core, not for individual rings.